+++ src/pages/dashboard/nutritionist/workshop/intro/index.tsx
@@ -1,5 +1,5 @@
 import DashboardOuterContainer from "@/common/components/DashboardOuterContainer";
-import NutritionistDashboardLayout from "@/modules.tsx/dashboard/nutritionist/DashboardLayout";
+import NutritionistDashboardLayout from "@/modules/dashboard/nutritionist/DashboardLayout";
 import NutritionistIntroForm from "@/modules/dashboard/nutritionist/workshop/NutritionistIntroForm";
 import Link from "next/link";
